requirements.txt
# --- Data & ML ---
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
nltk==3.9.1
matplotlib==3.10.1
seaborn==0.13.2

# --- Backend ---
fastapi==0.115.11
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1

# --- Database ---
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
alembic==1.15.1

# --- URL fetching ---
newspaper3k==0.2.8
lxml==5.2.2
lxml_html_clean==0.1.1

# --- Authentication ---
python-jose[cryptography]==3.3.0
#passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# --- Transformer model (GPU - Phase 6) ---
# Install PyTorch with CUDA first:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
transformers==4.49.0
datasets==3.3.2
accelerate==1.4.0

# --- Dev / Testing ---
pytest==8.3.5
httpx==0.28.1
anyio==4.9.0

bcrypt